+++ pkg/config/plugins/runtime/k8s/config.go
@@ -131,7 +131,7 @@ type KubernetesRuntimeConfig struct {
 type ClientConfig struct {
 	// Qps defines maximum requests kubernetes client is allowed to make per second.
 	// Default value 100. If set to 0 kube-client default value of 5 will be used.
-	Qps float32 `json:"qps" envconfig:"kuma_runtime_kubernetes_client_config_qps"`
+	Qps int `json:"qps" envconfig:"kuma_runtime_kubernetes_client_config_qps"`
 	// BurstQps defines maximum burst requests kubernetes client is allowed to make per second
 	// Default value 100. If set to 0 kube-client default value of 10 will be used.
 	BurstQps int `json:"burstQps" envconfig:"kuma_runtime_kubernetes_client_config_burst_qps"`
